v3python/tune/flash: renamed flash.py to module.py, reflecting the purpose
