+++ v3python/tune/flash/__init__.py
@@ -1,7 +1,7 @@
 # Copyright © 2025 Advanced Micro Devices, Inc.
 # SPDX-License-Identifier: MIT
 
-from .flash import (
+from .module import (
     Flash as TuneDesc,
     FlashKernelSelector as ImplSelector,
 )
